Rewrite of SAML2 HTTP Redirect request signing, now using xmlseclibs.
Implemented verification of authenticationrequest, logoutrequest and logoutresponse. 

Moved config of signing and verification to individual metadata entries. 

To sign/verify on hosted SP there must be a certificate and a privatekey. The normal certificate and privatekey of the IdP are reused when signing is enabled.

To verify remote SP and/or IdP the remote certificate must be present.

This commit is tested to some extend - please post bugs on google code project page.  


git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@178 44740490-163a-0410-bde0-09ae8108e29a