Session: Changed setData to have a default $timeout parameter.
This also changes the declaration of the setData method, as the $timeout
parameter moves to the end of the list.

If no $timeout parameter is supplied, then the 'session.datastore.timeout'
configuration option will be used. If this option doesn't exist, the
'session.requestcache' option will be used. If neither of these are used, then
four hours will be used as the timeout.


git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@576 44740490-163a-0410-bde0-09ae8108e29a