added support for IsPassive flag in Consent module and in SSOService.
If IsPassive is set and Consent requires user interaction an
SimpleSAML_Error_NoPassive exception is thrown. SSOService now
catches the exception and returns an AuthResponse with NoPassive
(which SSP currently can't handle in an orderly way).

We expect all processing modules to throw
SimpleSAML_Error_NoPassive exceptions if isPassive is true in
the state and they require user interaction.

ProcessingChain is updated with a new method -
processStatePassive which drives the processing filters with
isPassive set to true and ignores all SimpleSAML_Error_NoPassive
exceptions. This will be used in a forthcoming release of a
consent administration module.

Written by Jacob Christiansen and Mads Freek from wayf.dk.



git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@1237 44740490-163a-0410-bde0-09ae8108e29a