Added the wayf.dk consentAdmin functionality as a module. We have used this in wayf.dk for giving the users a way to see -, give - and withdraw consent for all service providers in one place. The work of adapting the code into a module mostly done by Jacob Christiansen - (jach@wayf.dk).
It depends on the Consent module for computing hashes and
accessing the consent database. It also depends on the new
isPassive functionality in the auth processing filters ie. a
filter must throw a SimpleSAML_Error_NoPassive exception if it
demands user interaction to be compatible with this module.

In addition Jacob has fixet a bug in the Consent module so that
i computes the correct targetedId when running on a bridge/proxy.

The database access in the Consent module has been updated with
some extra funtionality for consent administration use.



git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@1247 44740490-163a-0410-bde0-09ae8108e29a