_include.php: Improve handling of unhandled errors and exceptions.
This patch adds an exception handler which will catch all unhandled
exceptions and display an error page to the user. It also adds a
handler for PHP errors and warnings. This handler will show a stacktrace
of the error, and then pass the error to the normal PHP error handler.

git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@1561 44740490-163a-0410-bde0-09ae8108e29a