SAML: Work around namespace prefix limitation/bug in DOMXPath.
If a message contains a namespace prefix on the root node that conflicts
with the prefix used by simpleSAMLphp, XPath will use the original prefix.

This can be a problem if we receive a message with a namespace prefix
of 'saml' instead of 'samlp'.

See: http://bugs.php.net/bug.php?id=49490

git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@1747 44740490-163a-0410-bde0-09ae8108e29a