Also check $errno against PHP's error_reporting.
This is necessary to make sure that if a minimum PHP error level
has been set, this is also respected by SSP's error handler. Most
notably, this will make calls prefixed with '@', the PHP warning
silencer, actually silence the warnings. The SSP code tries to
makes use of the @ sign on various places around the code, but
this will only work if it's actually checked here.