Fix unnecessary writes with multiple memcache servers
Fixes a bug in Memcache.php that would cause every fetch to also
result in a push if multiple memcache servers were being used.

If multiple memcache servers are in use, the code needs to check
if they aren't out of sync. If they are, the most recent data needs
to be pushed to all mirrors.

The original code compared the parsed data structures using ===,
but that always fails because that just does a pointer comparison
and not a deep comparison.

Changed the code to compare the original unparsed values instead.
Also added a few debug-level log statements.