Set timeout when fetching metadata to 5 seconds.
When using the cron pattern with simplesamlphp, an http request kicks off the
cron job. Since we use FastCGI, the job is killed if it is not ready within the
maxtime set in fastcgi. This is ~90 seconds. If we get timeouts, the job is
killed. If the job is killed, all metadata is wiped and even servers that do
respond get their metadata wiped.  Pretty bad. So setting a low timeout makes
the probaility for this scenario much lower.