Ensure that config.php is optional for module loading.
Currently, if we try to load a class from a module, and the config.php
is absent, we will get an error about the missing configuration file.

There is no reason for the configuration file to be mandatory here,
since we have reasonable fallbacks, so fix the code to make it
optional.

This should make it easier to write tests for code in modules.