Show a user friendly error page
The authenticate.php file checks for the presence of the an exception within the state. If this is present the exception is formatted for logging and shown as plain text to the end user.

This has several disadvantages:
- The end user doesn't receive a user friendly error page.
- The exception is not logged within the authenticate.php
- Internal directory structure is exposed

For an example of such an situation/error see:
http://stackoverflow.com/questions/30511003/simplesamlphp-error

With the usage of SimpleSAML_exception_handler we'll solve the disadvantages.