Don't call `composer update` before Travis CI build
It looks like this was added in 855f4de961b7f2ba2b2415e7dac19a6128bfe52c
to avoid a dependency conflict -- the problem is that by calling
`composer update` before the build, you end up with a different set of
dependencies than is in `composer.lock`, meaning your test builds are
not operating in the same environment as "production".

This led to an issue with tests still passing even though the vfsStream
dependency was missing from `composer.lock` in #592.

If there are dependency conflicts in `composer.lock`, they should be
resolved locally, and an updated `composer.lock` pushed, or else others
will still deal with dependency issues.

This commit also sets the "php" version to 5.5.9 in the "platform" section
of the `composer.json` file, to foce dependencies to be compatible with
the minimum supported version of PHP.