Bugfix: searchformultiple() would stop on first dn even when nothing found.
searchformultiple() will loop over all the configured base dn's to
search in. However, it would break on the first search that does
not return `false`. A search that yields 0 hits is not `false`, however
(it is only false when an error occurred). So when using more than one
base, users would not be found if they were part of the second or later
base, which is contrary to the intention.

This is now changed so the loop breaks when the result is not false, and
the number of found results is > 0.