Show AuthData on authentication status page
It can be useful to see the AuthData when testing an authentication source.

Since the AuthData can be a hierarchy of data of various depths it is converted
into a pretty printed JSON structure.

Since the amount of AuthData can vary greatly, the data is displayed in side
a `<details>` element which requires clicking to view.