Add login.php and related twig template
Based on authenticate.php, login.php adds lines 6-26.
The script returns the possible authentication methods for the user, excluded admin.
When only one method is available the user is redirected to the one authentication-method login page.
Calls to right templates are updated at lines 17 and 54.

When no method is found, the user needs to login with admin credentials and check configuration.
The text shown in template is not translated.