docs: Spacing/indentation
The random line breaks everywhere in this file makes it hard to read at
narrow or wide window sizes whether you are using a Markdown reader or
just reading the markdown itself.

This puts all of the paragraphs and list items on their own line. The
one reading this file can decide for themselves how narrow or wide they
want the text to be displayed at.

I have also fixed the indentaiton of list items. Markdown will only read
a list item as "indented" when you use 4 spaces.

My final change in this commit was to remove all of the new lines below
headers. After reading so many markdown files, I find it much more
readable when there is one new line _before_ the header and zero new
lines below it.