Add authproc filter to test correct cardinality
Attribute schemas often have specific requirements for the number of
values a particular attribute may have, and violating the cardinality
rules associated with these may have unexpected and undesirable results.

The two authproc filters this change introduces allow the cardinality of
attributes to be tested. core:Cardinality is the more general case, and
allows for both a minimum and maximum number of values to be tested.
Processing will abort with an error if too few or too many are received.

core:CardinalitySingle is a special case of this for single-valued
attributes, and allows various corrective actions to be taken rather
than aborting. This allows for cardinality mistakes to be "fixed" in a
predictable way, and can also be used to construct single-valued
attributes from multi-valued ones.