Add attributes.exclude option to correspond with the Consent module
The Consent module has a (currently undocumented) noconsentattributes
option that allows specified attributes to be removed from the consent
hash calculation. The noconsentattribute option was introduced at
simplesamlphp/simplesamlphp@1efcfa8 but no corresponding option
was added to the consentAdmin module. Thus if people are using the
noconsentattribute option to remove attributes, the consentAdmin module
will *always* show this as attribute values having been changed.

This is a rework of simplesamlphp/simplesamlphp#531 which has gotten corrupted over time.