Update windowsliveid.md
In my search for using Microsoft as a claims provider I did not find this article, because I wasn't looking for Windows LiveID but an OIDC/oAuth2 replacement that fullfilled the userinfo endpoint using the Graph API. I hope the short introduction will help others find their way this invaluable the LiveID module. The wording may sound accusatory but I couldn't find a better way to describe the reasoning behind this module and create search engine keys to correctly identify it's purpose.