Add an asset() function to twig.
This function can be used to load an asset from a twig template, adding a "tag" parameter to it with a few dynamic bytes depending on the version of SimpleSAMLphp, or the last modification time of the file if we are running master. This behaves as a cache-buster, forcing browsers to reload assets when a new version of SimpleSAMLphp gets installed.