Actually fix the submit button in chrome.
The previous "fix" wasn't really fixing anything, it was indeed breaking the code that was changing the submit button, because jquery (and therefore "$") wasn't in scope for the script. Now, once jquery is available for that code, the bug appears again. This is due to an alleged "bug" in chrome, which prevents the button from actually working after it gets disabled. This makes full sense on the other hand, and as soon as we disable it, events won't keep being processed for that button, and as such, the form never gets submitted. This is why we need to submit the form manually.