Fix preferredidp in old UI disco 'links' style.
The calling code of the template had been changed to accommodate
new twig templates, but slightly changed the content of the
template variables so preferredidp never matched the list of known
idp's. This fixes it in the least invasive way. Could be
refactored because we now loop the idplist twice, but since this
is legacy code that's likely not worth it.

Thanks Stefan Winter from RESTENA for reporting.