Downgrade missing translations from error to debug
This is something the developer of a module is supposed to deal with. If you are just using SSP, it doesn't make much sense to have your log files flooded by errors about missing translation files or directories, since you cannot do anything about it and the original english text will be used anyway.